为仓库列表增加搜索查看功能，增加了货物列表

+++ backend/PMF/src/main/java/com/neuedu/pmf/service/impl/GoodsServiceImpl.java
@@ -1,6 +1,8 @@
 package com.neuedu.pmf.service.impl;
 
+import com.neuedu.pmf.common.ResultCode;
 import com.neuedu.pmf.entity.Goods;
+import com.neuedu.pmf.exception.BusinessException;
 import com.neuedu.pmf.mapper.GoodsMapper;
 import com.neuedu.pmf.service.GoodsService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -26,11 +28,31 @@ public class GoodsServiceImpl implements GoodsService {
 
     @Override
     public boolean delete(Integer id) {
+        Goods goods = goodsMapper.findById(id);
+        if (goods == null) {
+            throw new BusinessException(ResultCode.NOT_FOUND);
+        }
+        if (goods.getQuantity() != null && goods.getQuantity() > 0) {
+            throw new BusinessException(ResultCode.FAILED.getCode(), "该货物数量不为0，不能删除");
+        }
         return goodsMapper.deleteById(id) > 0;
     }
 
     @Override
     public boolean save(Goods goods) {
+        goods.setQuantity(0);
+
+        java.util.List<Integer> existingIds = goodsMapper.findAllIds();
+        int newId = 1;
+        for (int id : existingIds) {
+            if (id == newId) {
+                newId++;
+            } else if (id > newId) {
+                break;
+            }
+        }
+        goods.setGoods_id(newId);
+
         return goodsMapper.save(goods) > 0;
     }
